+++ erpc_c/setup/erpc_setup_mbox_zephyr.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 NXP
+ * Copyright 2023-2026 NXP
  *
  *
  * SPDX-License-Identifier: BSD-3-Clause
@@ -33,14 +33,14 @@ erpc_transport_t erpc_transport_zephyr_mbox_init(void *dev, void *tx_channel, vo
     }
     else
     {
-        s_transport.construct((struct device *)dev, (struct mbox_channel *)tx_channel,
-                              (struct mbox_channel *)rx_channel);
+        s_transport.construct((struct device *)dev, (struct mbox_dt_spec *)tx_channel,
+                              (struct mbox_dt_spec *)rx_channel);
         mboxTransport = s_transport.get();
     }
 
 #elif ERPC_ALLOCATION_POLICY == ERPC_ALLOCATION_POLICY_DYNAMIC
     mboxTransport =
-        new MBOXTransport((struct device *)dev, (struct mbox_channel *)tx_channel, (struct mbox_channel *)rx_channel);
+        new MBOXTransport((struct device *)dev, (struct mbox_dt_spec *)tx_channel, (struct mbox_dt_spec *)rx_channel);
 #else
 #error "Unknown eRPC allocation policy!"
 #endif